:recycle: refactor: enum, domain으로 통합
