--- src/main/java/com/example/harumeonglog/domain/event/entity/enums/EventCategory.java
@@ -1,9 +1,0 @@
-package com.example.harumeonglog.domain.event.entity.enums;
-
-public enum EventCategory {
-    BATH, // 목욕
-    WALK, // 산책
-    MEDICINE, // 약
-    MEDICAL, // 진료
-    ETC // 기타
-}
