--- src/main/java/com/example/harumeonglog/domain/member/entity/enums/NoticeType.java
@@ -1,7 +1,0 @@
-package com.example.harumeonglog.domain.member.entity.enums;
-
-public enum NoticeType {
-    EVENT, // 일정
-    ARTICLE, // 게시글
-    COMMENT // 댓글
-}
