--- src/main/java/com/example/harumeonglog/domain/member/entity/enums/SocialType.java
@@ -1,6 +1,0 @@
-package com.example.harumeonglog.domain.member.entity.enums;
-
-public enum SocialType {
-    KAKAO, // 카카오
-    APPLE // 애플
-}
