--- src/main/java/com/example/harumeonglog/domain/pet/entity/enums/Gender.java
@@ -1,6 +1,0 @@
-package com.example.harumeonglog.domain.pet.entity.enums;
-
-public enum Gender {
-    MALE, // 남
-    FEMALE // 여
-}
