--- src/main/java/com/example/harumeonglog/domain/pet/entity/enums/MemberPetRole.java
@@ -1,6 +1,0 @@
-package com.example.harumeonglog.domain.pet.entity.enums;
-
-public enum MemberPetRole {
-    OWNER, // Owner
-    GUEST // Guest
-}
