--- src/main/java/com/example/harumeonglog/domain/pet/entity/enums/PetSize.java
@@ -1,7 +1,0 @@
-package com.example.harumeonglog.domain.pet.entity.enums;
-
-public enum PetSize {
-    SMALL, // 소형견
-    MEDIUM, // 중형견
-    BIG // 대형견
-}
