--- src/main/java/com/example/harumeonglog/domain/post/controller/dto/enums/PostCategory.java
@@ -1,9 +1,0 @@
-package com.example.harumeonglog.domain.post.controller.dto.enums;
-
-public enum PostCategory {
-    INFO, // 정보 공유
-    HUMOR, // 유머
-    QNA, // 질문
-    SNS, // SNS
-    ETC // 기타
-}
